feat: Se cargaron validaciones de datos

+++ Colonia/src/main/java/Control/Colonia/Service/CasaServiceImplements.java
@@ -21,12 +21,34 @@ public class CasaServiceImplements implements CasaService {
 
     @Override
     public Casa saveCasa(Casa casa) throws RuntimeException {
+        try {
+            if (repo.existsByNoDeCasaAndDireccionAndEstadoAndPropietarioAndPrecioCasa(
+                    casa.getNoDeCasa(),
+                    casa.getDireccion(),
+                    casa.getEstado(),
+                    casa.getPropietario(),
+                    casa.getPrecioCasa())){
+                throw new RuntimeException("Ya existe una casa con estos datos");
+            }
+
             return repo.save(casa);
+        } catch (RuntimeException e) {
+            throw new RuntimeException(e.getMessage());
+        }
     }
 
     @Override
     public Casa updateCasa(Integer id, Casa casa) {
         Casa existingCasa = repo.findById(id).orElseThrow(() -> new RuntimeException("La casa no existe"));
+
+        if (repo.existsByNoDeCasaAndDireccionAndEstadoAndPropietarioAndPrecioCasa(
+                casa.getNoDeCasa(),
+                casa.getDireccion(),
+                casa.getEstado(),
+                casa.getPropietario(),
+                casa.getPrecioCasa())){
+            throw new RuntimeException("Ya existe una casa con estos datos");
+        }
 
         existingCasa.setNoDeCasa(casa.getNoDeCasa());
         existingCasa.setDireccion(casa.getDireccion());
@@ -39,6 +61,9 @@ public class CasaServiceImplements implements CasaService {
 
     @Override
     public void deleteCasa(Integer id) {
+        if(!repo.existsById(id)){
+            throw new RuntimeException("Este id no existe");
+        }
         repo.deleteById(id);
     }
 }
